refactor: move BlockType::as_func_type to validation
